+++ packages/sdk/src/trpc.ts
@@ -10,7 +10,7 @@ import {
 import SuperJSON from 'superjson'
 
 import type { OwnxTrpcRouter } from './api'
-import type { OwnxClientOptions } from './client';
+import type { OwnxClientOptions } from './client'
 import { makeHeaders } from './client'
 
 export type OwnxTrpcRouterInputs = inferRouterInputs<OwnxTrpcRouter>