+++ packages/chatbot/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ownxai/chatbot
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ownxai/sdk@0.1.6
+  - @ownxai/db@0.1.3
+  - @ownxai/tools@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes