+++ packages/db/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ownxai/db
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ownxai/shared@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes