+++ packages/mem0/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ownxai/mem0
 
+## 0.1.2
+
+### Patch Changes
+
+- @ownxai/db@0.1.3
+
 ## 0.1.1
 
 ### Patch Changes